pyproject.toml
[project]
name = "wjepa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.0.0",
    "numpy>=2.3.2",
    "pyyaml>=6.0.2",
    "soundfile>=0.13.1",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchcodec>=0.6.0",
    "tqdm>=4.67.1",
]
